--- a/packages/codegen/src/generateEntityCodegenFile.ts
+++ b/packages/codegen/src/generateEntityCodegenFile.ts
@@ -245,15 +245,6 @@ export function generateEntityCodegenFile(config: Config, dbMeta: DbMetadata, me
       static readonly tagName = "${tagName}";
       static readonly metadata: ${EntityMetadata}<${entity.type}>;
 
-      declare readonly __orm: {
-        entityType: ${entityName};
-        filterType: ${entityName}Filter;
-        gqlFilterType: ${entityName}GraphQLFilter;
-        orderType: ${entityName}Order;
-        fieldsType: ${entityName}Fields;
-        optIdsType: ${entityName}IdsOpts;
-        factoryOptsType: Parameters<typeof ${factoryMethod}>[1];
-      };     
       declare readonly __typeMapKeys: {
         ${baseEntity ? `0: "${baseEntity.name}",` : ""}
         ${baseEntity ? 1 : 0}: "${entityName}",