--- a/packages/orm/src/Entity.ts
+++ b/packages/orm/src/Entity.ts
@@ -36,4 +36,5 @@ export interface Entity {
    * name / display name.
    */
   toTaggedString(): string;
+  __typeMapKeys: unknown;
 }