--- a/packages/orm/src/EntityManager.ts
+++ b/packages/orm/src/EntityManager.ts
@@ -30,6 +30,7 @@ import {
   ExpressionFilter,
   FieldLogger,
   FilterWithAlias,
+  GraphQLFilterOf,
   GraphQLFilterWithAlias,
   InstanceData,
   Lens,
@@ -40,7 +41,6 @@ import {
   ReactionLogger,
   Reference,
   TimestampSerde,
-  TypeMap,
   UniqueFilter,
   ValidationError,
   ValidationErrors,
@@ -69,6 +69,7 @@ import { followReverseHint } from "./reactiveHints";
 import { ManyToOneReferenceImpl, OneToOneReferenceImpl, ReactiveReferenceImpl } from "./relations";
 import { AbstractRelationImpl } from "./relations/AbstractRelationImpl";
 import { AsyncMethodPopulateSecret } from "./relations/hasAsyncMethod";
+import { OptsOf, OrderOf } from "./typeMap";
 import { MaybePromise, assertNever, fail, failIfAnyRejected, getOrSet, groupBy, partition, toArray } from "./utils";
 
 // polyfill
@@ -128,60 +129,6 @@ export interface FindCountFilterOptions<T extends Entity> {
  * I.e. this is more like "MaybeAbstractEntityConstructor".
  */
 export type MaybeAbstractEntityConstructor<T> = abstract new (em: EntityManager<any, any>, opts: any) => T;
-
-/** Return the `FooOpts` type a given `Foo` entity constructor. */
-export type OptsOf<T> =
-  TypeMapKey<T> extends infer K ? (K extends keyof TypeMap ? TypeMap[K]["optsType"] : unknown) : unknown;
-
-/** Returns a key like `"Author"` or `"SmallPublisher"` for looking up types in the `TypeMap`. */
-type TypeMapKey<T> = T extends { __types: { 1: infer K } } ? K : T extends { __types: { 0: infer K } } ? K : never;
-
-// export type OptsOf<T> = T extends { __types: { 1: infer K } }
-//     ? K extends keyof TypeMap
-//         ? TypeMap[K]["optsType"]
-//         : unknown
-//     : T extends { __types: { 0: infer K } }
-//         ? K extends keyof TypeMap
-//             ? TypeMap[K]["optsType"]
-//             : unknown
-//         : unknown;
-//
-// export type OptsOf<T> = T extends { __orm: { optsType: infer O } } ? O : never;
-
-// export type OptsOf2<T extends { foo: string; typeMap: Record<string, { optsType: unknown }> }> =
-//   T["typeMap"][T["foo"]]["optsType"];
-
-export type FieldsOf<T> = T extends { __orm: { fieldsType: infer F } } ? F : never;
-
-export type OptIdsOf<T> = T extends { __orm: { optIdsType: infer O } } ? O : never;
-
-/** Return the `Foo` type for a given `Foo` entity constructor. */
-export type EntityOf<C> = C extends new (em: EntityManager, opts: any) => infer T ? T : never;
-
-/** Pulls the entity query type out of a given entity type T. */
-export type FilterOf<T> = T extends { __orm: { filterType: infer Q } } ? Q : never;
-
-/** Pulls the entity GraphQL query type out of a given entity type T. */
-export type GraphQLFilterOf<T> = T extends { __orm: { gqlFilterType: infer Q } } ? Q : never;
-
-/** Pulls the entity order type out of a given entity type T. */
-export type OrderOf<T> = T extends { __orm: { orderType: infer Q } } ? Q : never;
-
-type RelationsKeysOf<T> = {
-  [K in keyof OptsOf<T>]: OptsOf<T>[K] extends Entity[] | undefined ? K : never;
-}[keyof OptsOf<T>];
-
-/** Return the Relation keys from `FooOpt` type, given a `Foo` entity */
-export type RelationsOf<T extends Entity> = {
-  [K in RelationsKeysOf<T>]: NonNullable<OptsOf<T>[K]>;
-};
-
-/**
- * Returns the opts of the entity's `newEntity` factory method, as exists in the actual file.
- *
- * This is because `FactoryOpts` is a set of defaults, but the user can customize it if they want.
- */
-export type ActualFactoryOpts<T> = T extends { __orm: { factoryOptsType: infer Q } } ? Q : never;
 
 /** Pulls the entity's id type out of a given entity type T. */
 export type IdOf<T> = T extends { id: infer I } ? I : never;