--- a/packages/tests/esm-misc/src/entities/codegen/AuthorCodegen.ts
+++ b/packages/tests/esm-misc/src/entities/codegen/AuthorCodegen.ts
@@ -97,6 +97,21 @@ authorConfig.addRule(newRequiredRule("firstName"));
 authorConfig.addRule(newRequiredRule("createdAt"));
 authorConfig.addRule(newRequiredRule("updatedAt"));
 
+declare module "joist-orm" {
+  interface TypeMap {
+    Author: {
+      entityType: Author;
+      filterType: AuthorFilter;
+      gqlFilterType: AuthorGraphQLFilter;
+      orderType: AuthorOrder;
+      optsType: AuthorOpts;
+      fieldsType: AuthorFields;
+      optIdsType: AuthorIdsOpts;
+      factoryOptsType: Parameters<typeof newAuthor>[1];
+    };
+  }
+}
+
 export abstract class AuthorCodegen extends BaseEntity<EntityManager, string> implements Entity {
   static readonly tagName = "a";
   static readonly metadata: EntityMetadata<Author>;
@@ -106,11 +121,11 @@ export abstract class AuthorCodegen extends BaseEntity<EntityManager, string> im
     filterType: AuthorFilter;
     gqlFilterType: AuthorGraphQLFilter;
     orderType: AuthorOrder;
-    optsType: AuthorOpts;
     fieldsType: AuthorFields;
     optIdsType: AuthorIdsOpts;
     factoryOptsType: Parameters<typeof newAuthor>[1];
   };
+  declare readonly __types: { 0: "Author" };
 
   constructor(em: EntityManager, opts: AuthorOpts) {
     super(em, opts);