--- a/packages/tests/esm-misc/src/entities/codegen/AuthorCodegen.ts
+++ b/packages/tests/esm-misc/src/entities/codegen/AuthorCodegen.ts
@@ -116,15 +116,6 @@ export abstract class AuthorCodegen extends BaseEntity<EntityManager, string> im
   static readonly tagName = "a";
   static readonly metadata: EntityMetadata<Author>;
 
-  declare readonly __orm: {
-    entityType: Author;
-    filterType: AuthorFilter;
-    gqlFilterType: AuthorGraphQLFilter;
-    orderType: AuthorOrder;
-    fieldsType: AuthorFields;
-    optIdsType: AuthorIdsOpts;
-    factoryOptsType: Parameters<typeof newAuthor>[1];
-  };
   declare readonly __typeMapKeys: { 0: "Author" };
 
   constructor(em: EntityManager, opts: AuthorOpts) {