--- a/packages/tests/esm-misc/src/entities/codegen/AuthorCodegen.ts
+++ b/packages/tests/esm-misc/src/entities/codegen/AuthorCodegen.ts
@@ -116,7 +116,7 @@ export abstract class AuthorCodegen extends BaseEntity<EntityManager, string> im
   static readonly tagName = "a";
   static readonly metadata: EntityMetadata<Author>;
 
-  declare readonly __typeMapKeys: { 0: "Author" };
+  declare readonly __type: { 0: "Author" };
 
   constructor(em: EntityManager, opts: AuthorOpts) {
     super(em, opts);