Instead of using a hidden `__orm: { ...types... }` field to look up types like the `AuthorOpts` or `AuthorFilter`, we use a hidden `__type: { 0: "Author" } }` field that is "just a string" and then keep the types themselves in a global `TypeMap`, which are later "looked up" by the `OptsOf`, `FilterOf`, etc. helper types.

This keeps the `...Opts` / `...Filter` types from being a part of the entity's type signature, which is a simplification, and will hopefully give us some more freedom wrt filter/etc typing on subtypes (i.e. the filters of a subtype won't need to be co/contravariant with the filters of the base type).

Granted, we still need the `__type` field itself to do the lookups from "type `T`" --> "some other type", that given these "some other type"S are not directly derived/mapped from `T`.

(I.e. the other approach to this problem, is to rewrite all of the `AuthorOpts` / `AuthorFilter` / etc. types to purely be mapped types against `Author`; this is maybe doable, but codegen-ing out the types has so far being simpler to implement.)

Besides more freedom wrt filtering, the other hope was better compilation speeds, just from the entity types being (slightly) simpler. On our internal codebase, this change makes `tsc` ~6% faster, which isn't a crazy win, but still nice to see. 

Fixes #1295

Before:

```
$ time yarn build
________________________________________________________
Executed in   54.25 secs    fish           external
   usr time   72.73 secs  209.00 micros   72.73 secs
   sys time    2.35 secs   75.00 micros    2.35 secs
```

```
$ time yarn build
________________________________________________________
Executed in   53.43 secs    fish           external
   usr time   69.43 secs  276.00 micros   69.43 secs
   sys time    2.43 secs   99.00 micros    2.43 secs
```

```
$ time yarn build
________________________________________________________
Executed in   54.86 secs    fish           external
   usr time   73.87 secs  195.00 micros   73.87 secs
   sys time    2.49 secs   73.00 micros    2.49 secs
```

---

After

```
 $ time yarn build
________________________________________________________
Executed in   50.98 secs    fish           external
   usr time   66.98 secs  206.00 micros   66.98 secs
   sys time    2.52 secs   75.00 micros    2.52 secs
```

```
$ time yarn build
________________________________________________________
Executed in   51.56 secs    fish           external
   usr time   67.88 secs  192.00 micros   67.88 secs
   sys time    2.48 secs   70.00 micros    2.48 secs
```

```
$ time yarn build
________________________________________________________
Executed in   49.32 secs    fish           external
   usr time   65.03 secs  216.00 micros   65.03 secs
   sys time    2.40 secs   79.00 micros    2.40 secs
```